Address the comment from @jrhee17
